requirements.txt
requests
beautifulsoup4
logging
argparse
python-dotenv
